Upgrade to nightly-2023-06-15

+++ src/core/store/io/mmap_index_input.rs
@@ -178,7 +178,7 @@ impl MmapIndexInput {
             };
             let mmap = unsafe {
                 MmapOptions::new()
-                    .offset(offset)
+                    .offset(offset as u64)
                     .len(adapted_len)
                     .map(&file)?
             };
